+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Correctly handle second CTRL-C when waiting for worker threads.
 - Generate negative values for `minItems` & `maxItems` during the coverage phase.
 - Use `default` value as valid input during the coverage phase.
+- Unknown `multipart/form-data` fields not added to the final test case payload.
 
 ## [4.0.0-alpha.7](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.6...v4.0.0-alpha.7) - 2025-03-21